TagCloudPainter related fixes and refactoring
