+++ TagCloudCreator/Interfaces/IImageHolder.cs
@@ -1,6 +1,5 @@
 using System.Drawing;
 using TagCloudCreator.Domain.Settings;
-using TagCloudCreator.Domain;
 
 namespace TagCloudCreator.Interfaces;
