+++ TagCloudCreator/Interfaces/IWordsInfoParser.cs
@@ -1,9 +1,8 @@
 ﻿using TagCloudCreator.Infrastructure;
-using TagCloudCreator.Domain;
 
 namespace TagCloudCreator.Interfaces;
 
 public interface IWordsInfoParser
 {
-    IEnumerable<WordInfo> GetWordsInfos();
+    IEnumerable<WordInfo> GetWordsInfo();
 }
