+++ TagCloudCreator/Interfaces/Providers/IImageSettingsProvider.cs
@@ -1,5 +1,4 @@
 using TagCloudCreator.Domain.Settings;
-using TagCloudCreator.Domain;
 
 namespace TagCloudCreator.Interfaces.Providers;
