--- TagCloudCreator/Interfaces/Providers/IWordsInfosProvider.cs
@@ -1,9 +1,0 @@
-﻿using TagCloudCreator.Infrastructure;
-using TagCloudCreator.Domain;
-
-namespace TagCloudCreator.Interfaces.Providers;
-
-public interface IWordsInfosProvider
-{
-    IEnumerable<WordInfo> WordsInfos { get; }
-}
